Initial commit for full project structure and complete backend

+++ backend/controllers/contactControllers.js
@@ -4,27 +4,33 @@ const Contact = require('../models/contactModel');
 
 // @desc READ all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 // @desc READ contact
 // @route GET /api/contacts/:id
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error('Contact not found');
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to get other user's contact");
+    }
+
     res.status(200).json(contact);
 });
 
 // @desc CREATE contact
 // @route POST /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
     const {name,email,phone} = req.body;
     if(!name || !email || !phone){
@@ -36,7 +42,8 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id
     }) // as in es6 if the key and value are same, we can write it like this, instead of name: name, email: email, phone: phone
     res.status(201).json(contact);
     // console.log('Contact created: ', req.body);
@@ -44,16 +51,22 @@ const createContact = asyncHandler(async (req, res) => {
 
 // @desc UPDATE contact
 // @route PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
         res.status(404);
         throw new Error('Contact not found');
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to update other user's contact");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
-        req.params.id,
-        req.body,
+        req.params.id, //isse
+        req.body, // isme update kardo
         {new: true} // this will return the updated document instead of the old one
     );
     res.status(200).json(updatedContact);
@@ -61,13 +74,19 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc DELETE contact
 // @route DELETE /api/contacts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact) {
         res.status(404);
         throw new Error('Contact not found');
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to delete other user's contact");
+    }
+
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
 });
